fix: allow importing from browsers
When importing `@libp2p/tcp`, Instead of failing at compile time,
fail at runtime to make it easier to write isomorphic js.
