+++ packages/transport-tcp/.aegir.js
@@ -1,4 +1,5 @@
 
+/** @type {import('aegir').PartialOptions} */
 export default {
   build: {
     config: {
